refactor(review): decompose executor into responsibility-focused submodules
Deepens the #25 split beyond test extraction. Pure code movement,
no behavior change:

- types.rs: StepVerdict/StepOutcome/StepResult/RunOutcome/ExecutorError
- step.rs: single-step streaming with timeout/interrupt handling
- lifecycle.rs: bootstrap and teardown session exchanges
- checkpoint.rs: leading-checkpoint restore and post-step save
- transcript.rs: TranscriptWriter for incremental artifact capture
  and non-fatal artifact-error collection
- console.rs: step/summary progress rendering

mod.rs shrinks 810 -> 344 lines and now holds only the public API and
the orchestration loop; execute_steps_with_reporter drops from ~450 to
~250 lines by delegating to the modules above.
